Move things around and standardize syntax

+++ zsh/alias/tmux.zsh
@@ -1,6 +1,8 @@
-command -v tmux > /dev/null && {
+#!/usr/bin/env zsh
+
+if command -v tmux > /dev/null; then
   alias tls='tmux ls'
   alias tat='tmux attach -t'
   alias tnew='tmux new -s'
   alias tkill='tmux kill-session -t'
-}
+fi
